requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Environment Management
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1

# Password Hashing
bcrypt==4.1.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
requests==2.31.0

# Code Quality
pylint==3.0.3
black==23.12.1
flake8==7.0.0

# Utilities
python-dateutil==2.8.2